pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Daniel, Josh, Katie, Sam"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
requests = "^2.31.0"
pytest = "^8.0.2"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.1"
httpie = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
